Add test_document_foreign_access_denied_for_foreign_space
